Moved interfaces and types to separate .d.ts file
