+++ src/assets/_constants.scss
@@ -10,5 +10,5 @@ $ShadowBlack: #00000085;
 $ShadowGrey: #00000040;
 $TealBlue: #00b6ea;
 
-/* * * * URLS * * * */
+/* * * * URLs * * * */
 $BackgroundImage: "https://res.cloudinary.com/lenselink/image/upload/v1612725494/pubg-kill-feed/background.png";
