+++ tsconfig.json
@@ -14,7 +14,8 @@
     "baseUrl": "src",
 		"paths": {
       "Assets/*": ["assets/*"],
-      "Components/*": ["components/*"]
+      "Components/*": ["components/*"],
+			"*": ["components/Utils/*.d.ts"]
     }
 	},
 	"exclude": [
